+++ lib/src/config.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:flutter/widgets.dart';
 
 enum ColorMode {
